:bug: Fix(schema): 스키마 타입 에러 고침
거리두기 단계 스키마가 Int형으로 지정되서 전국 평균이 정상적으로 표시되지 않는 버그 수정

+++ src/schema/covid19-schema.ts
@@ -29,7 +29,7 @@ export const covid19Schema = `
     regionEng: String
     regionKor: String
     population: Int
-    distancingLevel: Int
+    distancingLevel: Float
     covid19DataList: [Covid19DataList]
   }
